requirements.txt
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.4.3
numpy>=1.21.0
gradio>=3.0.0
tqdm>=4.64.0
Pillow>=9.0.0
